refactor: rename `EditIconButton` to `EditEmojiIconButton`
