+++ apps/hash-frontend/src/shared/edit-emoji-icon-button/emoji-picker/emoji-picker.tsx
